requirements.txt
mysql-connector-python
bcrypt
cryptography
tabulate
